+++ internal/directus/login_test.go
@@ -2,13 +2,15 @@ package directus_test
 
 import (
 	"testing"
+
+	"github.com/Piitschy/drcts/test/testhelpers"
 )
 
 func TestGetToken(t *testing.T) {
-	ctx, container, d := NewDirectusContainer(t, "latest")
+	ctx, container, d := testhelpers.NewDirectusContainer(t, "latest")
 	defer container.Terminate(ctx)
 
-	auth, err := d.GetAuth(adminEmail, adminPassword)
+	auth, err := d.GetAuth(testhelpers.AdminEmail, testhelpers.AdminPassword)
 	if err != nil {
 		t.Fatalf("Failed to get token: %s", err)
 	}
@@ -18,7 +20,7 @@ func TestGetToken(t *testing.T) {
 }
 
 func TestGetTokenInvalid(t *testing.T) {
-	ctx, container, d := NewDirectusContainer(t, "latest")
+	ctx, container, d := testhelpers.NewDirectusContainer(t, "latest")
 	defer container.Terminate(ctx)
 
 	_, err := d.GetAuth("invalid", "invalid")
